Lab3/requirements.txt
alembic==1.13.1
greenlet==3.0.3
Mako==1.3.5
MarkupSafe==2.1.5
numpy==1.26.4
pandas==2.2.2
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.30
tabulate==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
